feat: Consolidate endpoint tests into a comprehensive script; remove outdated test file for new AI service endpoints
