--- microservices/ai-service/tests/test_new_endpoints.py
@@ -1,194 +1,0 @@
-#!/usr/bin/env python3
-"""
-Test script for the new AI service endpoints:
-- Hashtag Optimization
-- Optimal Posting Time Suggestion
-"""
-
-import requests
-import json
-import time
-
-# AI Service configuration
-AI_SERVICE_URL = "http://localhost:8005"
-
-def test_hashtag_optimization():
-    """Test the hashtag optimization endpoint"""
-    print("🧪 Testing Hashtag Optimization Endpoint...")
-    
-    url = f"{AI_SERVICE_URL}/api/ai/optimize/hashtags/"
-    
-    payload = {
-        "content": "Excited to announce our new AI-powered social media management platform! 🚀 We're helping businesses create engaging content, analyze performance, and optimize their social media strategy. Join us in revolutionizing how brands connect with their audience!",
-        "platform": "linkedin",
-        "target_audience": "professionals",
-        "industry": "technology"
-    }
-    
-    try:
-        response = requests.post(url, json=payload, timeout=30)
-        
-        print(f"Status Code: {response.status_code}")
-        
-        # Assertions for response
-        assert response.status_code in [200, 201], f"Expected 200/201, got {response.status_code}"
-        
-        if response.status_code in [200, 201]:
-            data = response.json()
-            print("✅ Hashtag Optimization Test PASSED")
-            print(f"Response: {json.dumps(data, indent=2)}")
-            
-            # Validate response structure with assertions
-            assert 'data' in data, "Response missing 'data' field"
-            response_data = data.get('data', {})
-            
-            if 'error' in response_data and response_data.get('fallback'):
-                print("⚠️  Using fallback response due to AI parsing issues")
-                print(f"Error: {response_data.get('error')}")
-                # Even fallback should have basic structure
-                assert 'hashtags' in response_data, "Fallback missing hashtags field"
-            elif 'hashtags' in response_data:
-                print(f"📊 Found {len(response_data['hashtags'])} hashtag suggestions")
-                assert isinstance(response_data['hashtags'], list), "Hashtags should be a list"
-                for hashtag in response_data['hashtags'][:3]:  # Show first 3
-                    print(f"  - {hashtag.get('tag', 'N/A')} ({hashtag.get('category', 'N/A')})")
-            else:
-                print("⚠️  Response structure may be incomplete")
-                assert False, "Response missing required hashtags field"
-                
-        else:
-            print(f"❌ Hashtag Optimization Test FAILED")
-            print(f"Error: {response.text}")
-            assert False, f"Request failed with status {response.status_code}"
-            
-    except Exception as e:
-        print(f"❌ Hashtag Optimization Test ERROR: {str(e)}")
-        assert False, f"Test failed with exception: {str(e)}"
-
-def test_optimal_posting_time():
-    """Test the optimal posting time endpoint"""
-    print("\n🧪 Testing Optimal Posting Time Endpoint...")
-    
-    url = f"{AI_SERVICE_URL}/api/ai/schedule/optimal/"
-    
-    payload = {
-        "platform": "instagram",
-        "content_type": "post",
-        "target_audience": "millennials",
-        "timezone": "America/New_York",
-        "industry": "fashion"
-    }
-    
-    try:
-        response = requests.post(url, json=payload, timeout=30)
-        
-        print(f"Status Code: {response.status_code}")
-        
-        # Assertions for response
-        assert response.status_code in [200, 201], f"Expected 200/201, got {response.status_code}"
-        
-        if response.status_code in [200, 201]:
-            data = response.json()
-            print("✅ Optimal Posting Time Test PASSED")
-            print(f"Response: {json.dumps(data, indent=2)}")
-            
-            # Validate response structure with assertions
-            assert 'data' in data, "Response missing 'data' field"
-            response_data = data.get('data', {})
-            
-            if 'error' in response_data and response_data.get('fallback'):
-                print("⚠️  Using fallback response due to AI parsing issues")
-                print(f"Error: {response_data.get('error')}")
-                # Even fallback should have basic structure
-                assert 'optimal_times' in response_data, "Fallback missing optimal_times field"
-            elif 'optimal_times' in response_data:
-                print(f"📅 Optimal times found for {len(response_data['optimal_times'])} days")
-                assert isinstance(response_data['optimal_times'], dict), "Optimal times should be a dict"
-                for day, times in list(response_data['optimal_times'].items())[:3]:  # Show first 3 days
-                    print(f"  - {day.capitalize()}: {', '.join(times)}")
-            else:
-                print("⚠️  Response structure may be incomplete")
-                assert False, "Response missing required optimal_times field"
-                
-        else:
-            print(f"❌ Optimal Posting Time Test FAILED")
-            print(f"Error: {response.text}")
-            assert False, f"Request failed with status {response.status_code}"
-            
-    except Exception as e:
-        print(f"❌ Optimal Posting Time Test ERROR: {str(e)}")
-        assert False, f"Test failed with exception: {str(e)}"
-
-def test_endpoint_comparison():
-    """Compare implemented endpoints with architecture requirements"""
-    print("\n📋 Endpoint Implementation Status:")
-    
-    required_endpoints = [
-        "POST /api/ai/generate/content",
-        "POST /api/ai/analyze/sentiment", 
-        "POST /api/ai/optimize/hashtags",
-        "POST /api/ai/schedule/optimal",
-        "GET /api/ai/models/status"
-    ]
-    
-    implemented_endpoints = [
-        "✅ POST /api/ai/generate/content",
-        "✅ POST /api/ai/analyze/sentiment",
-        "✅ POST /api/ai/optimize/hashtags",  # NEW
-        "✅ POST /api/ai/schedule/optimal",   # NEW
-        "✅ GET /api/ai/models/status"
-    ]
-    
-    for endpoint in implemented_endpoints:
-        print(f"  {endpoint}")
-    
-    print("\n🎉 All required AI service endpoints are now implemented!")
-
-def main():
-    """Run all tests"""
-    print("🚀 Testing New AI Service Endpoints")
-    print("=" * 50)
-    
-    test_results = []
-    
-    # Test the new endpoints
-    try:
-        test_hashtag_optimization()
-        test_results.append(True)
-    except AssertionError as e:
-        print(f"❌ Hashtag optimization test failed: {e}")
-        test_results.append(False)
-    except Exception as e:
-        print(f"❌ Hashtag optimization test error: {e}")
-        test_results.append(False)
-    
-    try:
-        test_optimal_posting_time()
-        test_results.append(True)
-    except AssertionError as e:
-        print(f"❌ Optimal posting time test failed: {e}")
-        test_results.append(False)
-    except Exception as e:
-        print(f"❌ Optimal posting time test error: {e}")
-        test_results.append(False)
-    
-    # Show implementation status
-    test_endpoint_comparison()
-    
-    # Summary
-    passed_tests = sum(test_results)
-    total_tests = len(test_results)
-    success_rate = (passed_tests / total_tests) * 100 if total_tests > 0 else 0
-    
-    print(f"\n📊 Test Results: {passed_tests}/{total_tests} tests passed ({success_rate:.1f}%)")
-    print("=" * 50)
-    
-    if passed_tests == total_tests:
-        print("✅ All tests passed!")
-    else:
-        print("❌ Some tests failed!")
-        import sys
-        sys.exit(1)
-
-if __name__ == "__main__":
-    main()
